+++ src/ambition/Ambition_input.cc
@@ -79,16 +79,14 @@ bool detectScroll(
   const PageSize horizontalPageSize,
   const PageSize horizontalMaximum
 ) {
-  if (enableMouse
-    && !area.contains({.left = mouse.cur_x, .top = mouse.cur_y,})
-  ) {
+  if (!area.contains({ .left = mouse.cur_x, .top = mouse.cur_y })) {
     return false;
   }
 
   int scrollHorizontal = 0;
   int scrollVertical = 0;
 
-  if (mouse.get_scroll(&scrollHorizontal, &scrollVertical)) {
+  if (enableMouse && mouse.get_scroll(&scrollHorizontal, &scrollVertical)) {
   } else {
     for (const auto& activation : activations) {
       int& scroll